+++ elastic-job-console/src/main/webapp/js/job_detail.js
@@ -10,6 +10,8 @@ $(function() {
     $('[href="#execution_info"]').click(function(event) {
         renderExecution();
     });
+    bindTriggerButtons();
+    bindTriggerAllButtons();
     bindPauseButtons();
     bindPauseAllButtons();
     bindResumeButtons();
@@ -22,6 +24,7 @@ $(function() {
 
 function renderSettings() {
     $.get("job/settings", {jobName : $("#job-name").text()}, function (data) {
+        $("#jobType").attr("value", data.jobType);
         $("#jobClass").attr("value", data.jobClass);
         $("#shardingTotalCount").attr("value", data.shardingTotalCount);
         $("#cron").attr("value", data.cron);
@@ -33,6 +36,7 @@ function renderSettings() {
         $("#processCountIntervalSeconds").attr("value", data.processCountIntervalSeconds);
         $("#concurrentDataProcessThreadCount").attr("value", data.concurrentDataProcessThreadCount);
         $("#fetchDataCount").attr("value", data.fetchDataCount);
+        $("#streamingProcess").attr("checked", data.streamingProcess);
         $("#maxTimeDiffSeconds").attr("value", data.maxTimeDiffSeconds);
         $("#monitorPort").attr("value", data.monitorPort);
         $("#jobShardingStrategyClass").attr("value", data.jobShardingStrategyClass);
@@ -40,6 +44,7 @@ function renderSettings() {
         if (!data.monitorExecution) {
             $("#execution_info_tab").addClass("disabled");
         }
+        $("#scriptCommandLine").attr("value", data.scriptCommandLine);
     });
 }
 
@@ -47,6 +52,7 @@ function bindSubmitJobSettingsForm() {
     $("#job-settings-form").submit(function(event) {
         event.preventDefault();
         var jobName = $("#job-name").text();
+        var jobType = $("#jobType").val();
         var jobClass = $("#jobClass").val();
         var shardingTotalCount = $("#shardingTotalCount").val();
         var jobParameter = $("#jobParameter").val();
@@ -54,6 +60,7 @@ function bindSubmitJobSettingsForm() {
         var concurrentDataProcessThreadCount = $("#concurrentDataProcessThreadCount").val();
         var processCountIntervalSeconds = $("#processCountIntervalSeconds").val();
         var fetchDataCount = $("#fetchDataCount").val();
+        var streamingProcess = $("#streamingProcess").prop("checked");
         var maxTimeDiffSeconds = $("#maxTimeDiffSeconds").val();
         var monitorPort = $("#monitorPort").val();
         var monitorExecution = $("#monitorExecution").prop("checked");
@@ -61,8 +68,9 @@ function bindSubmitJobSettingsForm() {
         var misfire = $("#misfire").prop("checked");
         var shardingItemParameters = $("#shardingItemParameters").val();
         var jobShardingStrategyClass = $("#jobShardingStrategyClass").val();
+        var scriptCommandLine = $("#scriptCommandLine").val();
         var description = $("#description").val();
-        $.post("job/settings", {jobName: jobName, jobClass : jobClass, shardingTotalCount: shardingTotalCount, jobParameter: jobParameter, cron: cron, concurrentDataProcessThreadCount: concurrentDataProcessThreadCount, processCountIntervalSeconds: processCountIntervalSeconds, fetchDataCount: fetchDataCount, maxTimeDiffSeconds: maxTimeDiffSeconds, monitorPort: monitorPort, monitorExecution: monitorExecution, failover: failover, misfire: misfire, shardingItemParameters: shardingItemParameters, jobShardingStrategyClass: jobShardingStrategyClass, description: description}, function(data) {
+        $.post("job/settings", {jobName: jobName, jobType : jobType, jobClass : jobClass, shardingTotalCount: shardingTotalCount, jobParameter: jobParameter, cron: cron, concurrentDataProcessThreadCount: concurrentDataProcessThreadCount, processCountIntervalSeconds: processCountIntervalSeconds, fetchDataCount: fetchDataCount, streamingProcess: streamingProcess, maxTimeDiffSeconds: maxTimeDiffSeconds, monitorPort: monitorPort, monitorExecution: monitorExecution, failover: failover, misfire: misfire, shardingItemParameters: shardingItemParameters, jobShardingStrategyClass: jobShardingStrategyClass, description: description, scriptCommandLine: scriptCommandLine}, function(data) {
             showSuccessDialog();
             if (monitorExecution) {
                 $("#execution_info_tab").removeClass("disabled");
@@ -80,16 +88,18 @@ function renderServers() {
             var status = data[i].status;
             var baseTd = "<td>" + data[i].ip + "</td><td>" + data[i].hostName + "</td><td>" + status + "</td><td>" + data[i].processSuccessCount + "</td><td>" + data[i].processFailureCount + "</td><td>" + data[i].sharding + "</td>";
             var operationTd = "";
+            var triggerButton = "<button operation='trigger' class='btn btn-success' ip='" + data[i].ip + "'>触发</button>";
             var resumeButton = "<button operation='resume' class='btn btn-success' ip='" + data[i].ip + "'>恢复</button>";
             var pauseButton = "<button operation='pause' class='btn btn-warning' ip='" + data[i].ip + "'" + ">暂停</button>";
             var shutdownButton = "<button operation='shutdown' class='btn btn-danger' ip='" + data[i].ip + "'>关闭</button>";
             var removeButton = "<button operation='remove' class='btn btn-danger' ip='" + data[i].ip + "'>删除</button>";
             var disableButton = "<button operation='disable' class='btn btn-danger' ip='" + data[i].ip + "'>失效</button>";
             var enableButton = "<button operation='enable' class='btn btn-success' ip='" + data[i].ip + "'>生效</button>";
+            operationTd = triggerButton + "&nbsp;";
             if ("PAUSED" === status) {
-                operationTd = resumeButton + "&nbsp;";
+                operationTd = operationTd + resumeButton + "&nbsp;";
             } else if ("DISABLED" !== status && "CRASHED" !== status && "SHUTDOWN" !== status) {
-                operationTd = pauseButton + "&nbsp;";
+                operationTd = operationTd + pauseButton + "&nbsp;";
             }
             if ("SHUTDOWN" !== status) {
                 operationTd = operationTd + shutdownButton + "&nbsp;";
@@ -115,6 +125,26 @@ function renderServers() {
             }
             $("#servers tbody").append("<tr class='" + trClass + "'>" + baseTd + operationTd + "</tr>");
         }
+    });
+}
+
+function bindTriggerButtons() {
+    $(document).on("click", "button[operation='trigger'][data-toggle!='modal']", function(event) {
+        var jobName = $("#job-name").text();
+        $.post("job/trigger", {jobName : jobName, ip : $(event.currentTarget).attr("ip")}, function (data) {
+            renderServers();
+            showSuccessDialog();
+        });
+    });
+}
+
+function bindTriggerAllButtons() {
+    $(document).on("click", "#trigger-all-jobs-btn", function(event) {
+        var jobName = $("#job-name").text();
+        $.post("job/triggerAll/name", {jobName : jobName}, function (data) {
+            renderServers();
+            showSuccessDialog();
+        });
     });
 }