+++ elastic-job-core/src/test/java/com/dangdang/ddframe/job/internal/server/JobOperationListenerManagerTest.java
@@ -17,14 +17,17 @@
 
 package com.dangdang.ddframe.job.internal.server;
 
-import static org.mockito.Mockito.times;
-import static org.mockito.Mockito.verify;
-import static org.mockito.Mockito.when;
-
+import com.dangdang.ddframe.job.api.config.JobConfigurationFactory;
+import com.dangdang.ddframe.job.fixture.TestJob;
 import com.dangdang.ddframe.job.internal.election.LeaderElectionService;
+import com.dangdang.ddframe.job.internal.env.LocalHostService;
 import com.dangdang.ddframe.job.internal.execution.ExecutionService;
+import com.dangdang.ddframe.job.internal.schedule.JobRegistry;
 import com.dangdang.ddframe.job.internal.schedule.JobScheduleController;
+import com.dangdang.ddframe.job.internal.server.JobOperationListenerManager.ConnectionLostListener;
+import com.dangdang.ddframe.job.internal.server.JobOperationListenerManager.JobPausedStatusJobListener;
 import com.dangdang.ddframe.job.internal.sharding.ShardingService;
+import com.dangdang.ddframe.job.internal.storage.JobNodeStorage;
 import org.apache.curator.framework.recipes.cache.ChildData;
 import org.apache.curator.framework.recipes.cache.TreeCacheEvent;
 import org.apache.curator.framework.state.ConnectionState;
@@ -35,15 +38,11 @@ import org.mockito.Mock;
 import org.mockito.MockitoAnnotations;
 import org.unitils.util.ReflectionUtils;
 
-import com.dangdang.ddframe.job.api.JobConfiguration;
-import com.dangdang.ddframe.job.fixture.TestJob;
-import com.dangdang.ddframe.job.internal.env.LocalHostService;
-import com.dangdang.ddframe.job.internal.schedule.JobRegistry;
-import com.dangdang.ddframe.job.internal.server.JobOperationListenerManager.ConnectionLostListener;
-import com.dangdang.ddframe.job.internal.server.JobOperationListenerManager.JobPausedStatusJobListener;
-import com.dangdang.ddframe.job.internal.storage.JobNodeStorage;
-
 import java.util.Arrays;
+
+import static org.mockito.Mockito.times;
+import static org.mockito.Mockito.verify;
+import static org.mockito.Mockito.when;
 
 public final class JobOperationListenerManagerTest {
     
@@ -67,7 +66,8 @@ public final class JobOperationListenerManagerTest {
     
     private String ip = new LocalHostService().getIp();
     
-    private final JobOperationListenerManager jobOperationListenerManager = new JobOperationListenerManager(null, new JobConfiguration("testJob", TestJob.class, 3, "0/1 * * * * ?"));
+    private final JobOperationListenerManager jobOperationListenerManager = new JobOperationListenerManager(null, 
+            JobConfigurationFactory.createSimpleJobConfigurationBuilder("testJob", TestJob.class, 3, "0/1 * * * * ?").build());
     
     @Before
     public void setUp() throws NoSuchFieldException {
@@ -83,7 +83,7 @@ public final class JobOperationListenerManagerTest {
     public void assertStart() {
         jobOperationListenerManager.start();
         verify(jobNodeStorage).addConnectionStateListener(Matchers.<ConnectionLostListener>any());
-        verify(jobNodeStorage, times(2)).addDataListener(Matchers.<JobPausedStatusJobListener>any());
+        verify(jobNodeStorage, times(3)).addDataListener(Matchers.<JobPausedStatusJobListener>any());
     }
     
     @Test
@@ -121,6 +121,51 @@ public final class JobOperationListenerManagerTest {
         jobOperationListenerManager.new ConnectionLostListener().stateChanged(null, ConnectionState.CONNECTED);
         verify(jobScheduleController, times(0)).pauseJob();
         verify(jobScheduleController, times(0)).resumeJob();
+    }
+    
+    @Test
+    public void assertJobTriggerStatusJobListenerWhenRemove() {
+        jobOperationListenerManager.new JobTriggerStatusJobListener().dataChanged(null, new TreeCacheEvent(
+                TreeCacheEvent.Type.NODE_REMOVED, new ChildData("/testJob/servers/" + ip + "/trigger", null, "".getBytes())), "/testJob/servers/" + ip + "/trigger");
+        verify(serverService, times(0)).clearJobTriggerStatus();
+        verify(jobScheduleController, times(0)).triggerJob();
+    }
+    
+    @Test
+    public void assertJobTriggerStatusJobListenerWhenIsAddButNotLocalHostJobTriggerPath() {
+        jobOperationListenerManager.new JobTriggerStatusJobListener().dataChanged(null, new TreeCacheEvent(
+                TreeCacheEvent.Type.NODE_ADDED, new ChildData("/testJob/servers/" + ip + "/other", null, "".getBytes())), "/testJob/servers/" + ip + "/other");
+        verify(serverService, times(0)).clearJobTriggerStatus();
+        verify(jobScheduleController, times(0)).triggerJob();
+    }
+    
+    @Test
+    public void assertJobTriggerStatusJobListenerWhenIsAddAndIsJobLocalHostTriggerPathButNoJobRegister() {
+        jobOperationListenerManager.new JobTriggerStatusJobListener().dataChanged(null, new TreeCacheEvent(
+                TreeCacheEvent.Type.NODE_ADDED, new ChildData("/testJob/servers/" + ip + "/trigger", null, "".getBytes())), "/testJob/servers/" + ip + "/trigger");
+        verify(serverService).clearJobTriggerStatus();
+        verify(jobScheduleController, times(0)).triggerJob();
+    }
+    
+    @Test
+    public void assertJobTriggerStatusJobListenerWhenIsAddAndIsJobLocalHostTriggerPathAndJobRegisterButServerIsNotReady() {
+        JobRegistry.getInstance().addJobScheduleController("testJob", jobScheduleController);
+        jobOperationListenerManager.new JobTriggerStatusJobListener().dataChanged(null, new TreeCacheEvent(
+                TreeCacheEvent.Type.NODE_ADDED, new ChildData("/testJob/servers/" + ip + "/trigger", null, "".getBytes())), "/testJob/servers/" + ip + "/trigger");
+        verify(serverService).clearJobTriggerStatus();
+        verify(serverService).isLocalhostServerReady();
+        verify(jobScheduleController, times(0)).triggerJob();
+    }
+    
+    @Test
+    public void assertJobTriggerStatusJobListenerWhenIsAddAndIsJobLocalHostTriggerPathAndJobRegisterAndServerIsReady() {
+        JobRegistry.getInstance().addJobScheduleController("testJob", jobScheduleController);
+        when(serverService.isLocalhostServerReady()).thenReturn(true);
+        jobOperationListenerManager.new JobTriggerStatusJobListener().dataChanged(null, new TreeCacheEvent(
+                TreeCacheEvent.Type.NODE_ADDED, new ChildData("/testJob/servers/" + ip + "/trigger", null, "".getBytes())), "/testJob/servers/" + ip + "/trigger");
+        verify(serverService).isLocalhostServerReady();
+        verify(jobScheduleController).triggerJob();
+        verify(serverService).clearJobTriggerStatus();
     }
     
     @Test