+++ elastic-job-core/src/test/java/com/dangdang/ddframe/job/internal/statistics/ProcessCountJobTest.java
@@ -17,10 +17,9 @@
 
 package com.dangdang.ddframe.job.internal.statistics;
 
-import static org.hamcrest.CoreMatchers.is;
-import static org.junit.Assert.assertThat;
-import static org.mockito.Mockito.verify;
-
+import com.dangdang.ddframe.job.api.config.JobConfigurationFactory;
+import com.dangdang.ddframe.job.fixture.TestJob;
+import com.dangdang.ddframe.job.internal.server.ServerService;
 import org.junit.Before;
 import org.junit.Test;
 import org.mockito.Mock;
@@ -28,16 +27,18 @@ import org.mockito.MockitoAnnotations;
 import org.quartz.JobExecutionException;
 import org.unitils.util.ReflectionUtils;
 
-import com.dangdang.ddframe.job.api.JobConfiguration;
-import com.dangdang.ddframe.job.fixture.TestJob;
-import com.dangdang.ddframe.job.internal.server.ServerService;
+import static org.hamcrest.CoreMatchers.is;
+import static org.junit.Assert.assertThat;
+import static org.mockito.Mockito.verify;
 
 public final class ProcessCountJobTest {
     
     @Mock
     private ServerService serverService;
     
-    private final ProcessCountJob processCountJob = new ProcessCountJob(null, new JobConfiguration(ProcessCountJobTest.class.getName(), TestJob.class, 3, "0/1 * * * * ?"));
+    private final ProcessCountJob processCountJob = new ProcessCountJob(null, 
+            JobConfigurationFactory.createSimpleJobConfigurationBuilder(ProcessCountJobTest.class.getName(), 
+                    TestJob.class, 3, "0/1 * * * * ?").build());
     
     @Before
     public void setUp() throws NoSuchFieldException {